backend/requirements.txt
azure-identity==1.13.0
quart==0.18.4
langchain==0.0.187
openai[datalib]==0.27.8
tiktoken==0.4.0
azure-search-documents==11.4.0b6
azure-storage-blob==12.14.1
azure-cosmos==4.5.0
uvicorn[standard]==0.23.2
Werkzeug==2.3.7
opencensus-ext-azure==1.1.9
azure-monitor-opentelemetry==1.0.0b15
opentelemetry-instrumentation-flask==0.40b0